hack_retr/requirements.txt
transformers==4.35.2
python-dotenv==1.0.0
fastapi==0.109.2
pydantic==2.6.1
uvicorn==0.27.1
pillow
sentence-transformers==2.2.2
chromadb